+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Date format: DD/MM/YYYY
 
 ## [2.0.0] - [07/05/2021]
 
+- New way to disable the acrylic blur effect. Just wrap the acrylic widget in a `NoAcrylicBlurEffect` to have it disabled.
 
 ## [1.10.1] - [05/05/2021]